# controller/role.go
package controller

import (
	"github.com/alert666/api-server/base/bind"
	v1 "github.com/alert666/api-server/service/v1"
	"github.com/gin-gonic/gin"
)

type RoleController interface {
	CreateRole(c *gin.Context)
	UpdateRole(c *gin.Context)
	DeleteRole(c *gin.Context)
	QueryRole(c *gin.Context)
	ListRole(c *gin.Context)
}

type roleController struct {
	roleService v1.RoleServicer
}

func NewRoleController(roleService v1.RoleServicer) RoleController {
	return &roleController{
		roleService: roleService,
	}
}

// CreateRole 创建角色
// @Summary 创建角色
// @Description 创建角色
// @Tags 角色管理
// @Accept json
// @Produce json
// @Param data body types.RoleCreateRequest true "创建请求参数"
// @Success 200 {object} types.Response "创建成功"
// @Router /api/v1/role [post]
func (receiver *roleController) CreateRole(c *gin.Context) {
	bind.ResponseOnlySuccess(c, receiver.roleService.CreateRole, bind.BindTypeJson)
}

// UpdateRole 更新角色
// @Summary 更新角色
// @Description 更新角色, 并且可以更新角色的权限
// @Tags 角色管理
// @Accept json
// @Produce json
// @Param data body types.RoleUpdateRequest true "更新请求参数"
// @Success 200 {object} types.Response "更新成功"
// @Router /api/v1/role/:id [put]
func (receiver *roleController) UpdateRole(c *gin.Context) {
	bind.ResponseOnlySuccess(c, receiver.roleService.UpdateRole, bind.BindTypeUri, bind.BindTypeJson)
}

// DeleteRole 删除角色
// @Summary 删除角色
// @Description 删除角色, 不能删除有用户的角色
// @Tags 角色管理
// @Accept json
// @Produce json
// @Param data body types.IDRequest true "删除请求参数"
// @Success 200 {object} types.Response "删除成功"
// @Router /api/v1/role/:id [delete]
func (receiver *roleController) DeleteRole(c *gin.Context) {
	bind.ResponseOnlySuccess(c, receiver.roleService.DeleteRole, bind.BindTypeUri)
}

// QueryRole 查询角色
// @Summary 查询角色
// @Description 查询角色, 包括角色的权限
// @Tags 角色管理
// @Accept json
// @Produce json
// @Param data body types.IDRequest true "查询请求参数"
// @Success 200 {object} types.Response{} "查询成功"
// @Router /api/v1/role/:id [get]
func (receiver *roleController) QueryRole(c *gin.Context) {
	bind.ResponseWithData(c, receiver.roleService.QueryRole, bind.BindTypeUri)
}

// ListRole 角色列表
// @Summary 角色列表
// @Description 使用分页查询角色的信息, 支持根据 name 查询
// @Tags 角色管理
// @Accept json
// @Produce json
// @Param data query types.RoleListRequest true "查询请求参数"
// @Success 200 {object} types.Response{} "查询成功"
// @Router /api/v1/role [get]
func (receiver *roleController) ListRole(c *gin.Context) {
	bind.ResponseWithData(c, receiver.roleService.ListRole, bind.BindTypeUri, bind.BindTypeQuery)
}